Refactor: isolate decision receipt projection
