+++ src/audit/execution-identity-context.ts
@@ -3,14 +3,10 @@ import type { DatabaseSync } from "node:sqlite";
 import type { Selectable } from "kysely";
 import type {
   AuditRunInspectResult,
-  DecisionReceiptV1,
   ExecutionIdentityContextV1,
 } from "../../packages/gateway-protocol/src/index.js";
 import { validateExecutionIdentityContextV1 } from "../../packages/gateway-protocol/src/index.js";
-import {
-  countOperatorApprovalReceiptsForRun,
-  listOperatorApprovalReceiptsForRun,
-} from "../gateway/operator-approval-store.js";
+import { countOperatorApprovalReceiptsForRun } from "../gateway/operator-approval-store.js";
 import {
   executeSqliteQuerySync,
   executeSqliteQueryTakeFirstSync,
@@ -26,11 +22,8 @@ import {
   type OpenClawStateDatabaseOptions,
 } from "../state/openclaw-state-db.js";
 import { clearAuditIdentityKeyCacheForDatabase } from "./audit-identity.js";
-import {
-  countExecutionDecisionFactsForContext,
-  countExecutionDecisionFactsForRun,
-  listExecutionDecisionFactsForContext,
-} from "./execution-decision-facts.js";
+import { countExecutionDecisionFactsForRun } from "./execution-decision-facts.js";
+import { presentExecutionDecisionReceipts } from "./execution-decision-receipts.js";
 import {
   parseExecutionIdentityAdmissionEnvelope,
   parseExecutionIdentityAdmissionWork,
@@ -390,44 +383,6 @@ function readExecutionIdentityContextByExecutionId(
   );
 }
 
-function admissionDecision(context: ExecutionIdentityContextV1): DecisionReceiptV1 {
-  return {
-    schemaVersion: 1,
-    receiptId: `${context.contextId}:admission`,
-    contextId: context.contextId,
-    executionId: context.executionId,
-    runId: context.runId,
-    occurredAt: context.createdAt,
-    action: {
-      family: "run",
-      operation: "admission",
-      summary: "Run admission was recorded without an identity-aware policy or grant decision.",
-    },
-    decision: {
-      outcome: "not-applicable",
-      reasonCode: "run_admission_identity_not_evaluated",
-    },
-    enforcement: {
-      coverageState: context.coverageState,
-      policyRefs: [],
-      grantRefs: [],
-      contextFieldsUsed: [],
-    },
-    source: {
-      owner: "agent-command",
-      recordRef: context.contextId,
-      decisionBoundary: "agent-command.run-admission",
-    },
-    missingEvidence: [...context.missingEvidence],
-    remediation: [
-      {
-        code: "no_identity_enforcement_claimed",
-        text: "Treat this receipt as attribution only; it does not prove authorization.",
-      },
-    ],
-  };
-}
-
 function unavailableResult(params: {
   selector: { runId: string } | { executionId: string };
   resolvedRunId?: string;
@@ -478,97 +433,6 @@ function unavailableIdentityContext(
   });
 }
 
-function presentResult(params: {
-  context: ExecutionIdentityContextV1;
-  decisionOffset?: number;
-  decisionLimit?: number;
-  options: ExecutionIdentityReadOptions;
-}): AuditRunInspectResult {
-  const offset = params.decisionOffset ?? 0;
-  const limit = params.decisionLimit ?? 50;
-  const now = params.options.now ?? Date.now();
-  const approvalCount = countOperatorApprovalReceiptsForRun({
-    runId: params.context.runId,
-    nowMs: now,
-    databaseOptions: params.options,
-  });
-  const genericCount = countExecutionDecisionFactsForContext({
-    contextId: params.context.contextId,
-    now,
-    database: params.options,
-  });
-  const totalDecisions = 1 + approvalCount + genericCount;
-  const decisions: DecisionReceiptV1[] = [];
-  let remainingOffset = offset;
-  let remainingLimit = limit;
-
-  if (remainingOffset === 0 && remainingLimit > 0) {
-    decisions.push(admissionDecision(params.context));
-    remainingLimit -= 1;
-  } else {
-    remainingOffset = Math.max(0, remainingOffset - 1);
-  }
-  if (remainingLimit > 0 && remainingOffset < approvalCount) {
-    const page = listOperatorApprovalReceiptsForRun({
-      context: {
-        contextId: params.context.contextId,
-        executionId: params.context.executionId,
-        runId: params.context.runId,
-      },
-      offset: remainingOffset,
-      limit: remainingLimit,
-      nowMs: now,
-      databaseOptions: params.options,
-    });
-    decisions.push(...page);
-    remainingLimit -= page.length;
-    remainingOffset = 0;
-  } else {
-    remainingOffset = Math.max(0, remainingOffset - approvalCount);
-  }
-  if (remainingLimit > 0 && remainingOffset < genericCount) {
-    decisions.push(
-      ...listExecutionDecisionFactsForContext({
-        contextId: params.context.contextId,
-        offset: remainingOffset,
-        limit: remainingLimit,
-        now,
-        database: params.options,
-      }),
-    );
-  }
-  const nextOffset = offset + decisions.length;
-  const pageCoverage = decisions.map((receipt) => receipt.enforcement.coverageState);
-  const coverageState = pageCoverage.includes("unsupported")
-    ? "unsupported"
-    : pageCoverage.includes("unknown")
-      ? "unknown"
-      : approvalCount > 0 || pageCoverage.includes("enforced")
-        ? "enforced"
-        : params.context.coverageState;
-  const missingEvidence = [
-    ...new Set([
-      ...params.context.missingEvidence,
-      ...decisions.flatMap((receipt) => receipt.missingEvidence),
-    ]),
-  ].toSorted();
-  return {
-    schemaVersion: 1,
-    run: {
-      runId: params.context.runId,
-      executionId: params.context.executionId,
-      status: "known",
-    },
-    identity: { state: "present", context: params.context },
-    decisions,
-    coverage: {
-      state: coverageState,
-      missingEvidence,
-    },
-    ...(nextOffset < totalDecisions ? { nextDecisionCursor: String(nextOffset) } : {}),
-  };
-}
-
 function inspectExactExecution(
   params: { executionId: string; decisionOffset?: number; decisionLimit?: number },
   options: ExecutionIdentityReadOptions,
@@ -577,7 +441,7 @@ function inspectExactExecution(
   const selector = { executionId };
   const contextResult = readExecutionIdentityContextByExecutionId(executionId, options);
   if (contextResult.status === "found") {
-    return presentResult({
+    return presentExecutionDecisionReceipts({
       context: contextResult.context,
       decisionOffset: params.decisionOffset,
       decisionLimit: params.decisionLimit,
@@ -675,7 +539,7 @@ function inspectRunSelector(
         : [];
       if (firstMatches.length === 1) {
         try {
-          return presentResult({
+          return presentExecutionDecisionReceipts({
             context: parseExecutionIdentityRow(firstMatches[0]!),
             decisionOffset: params.decisionOffset,
             decisionLimit: params.decisionLimit,
